Revert "Merge branch 'fixSelectMain' of https://github.com/antonymofyou/HireEmployeesVue into fixSelectMain"
This reverts commit 4787b8fd93f99545e057a44e4e5da2a6582f64cf, reversing
changes made to 6ce6f2614c85a92a7ac4178851b6923a5e7cc74b.
