+++ src/components/SelectMain.vue
@@ -34,7 +34,7 @@
 </template>
 
 <script setup>
-import { ref, computed, watchEffect } from 'vue';
+import { ref, computed, onMounted, watch } from 'vue';
 
 const props = defineProps({
   // Модель для обновления, опции селекта
@@ -84,9 +84,14 @@ const optionsContainerStyle = computed(() => {
   };
 });
 
-// Отслеживание изменений, отработает и при монтировании
-watchEffect(() => {
+// Установка значения и орбаботчика при монтировании
+onMounted(() => {
   selected.value = props.options.find(option => option.id == props.modelValue) || null;
+});
+
+// Отслеживание изменений
+watch(() => props.modelValue, (newValue) => {
+  selected.value = props.options.find(option => option.id == newValue) || null;
 });
 
 // Директива для закрытия селекта по клику снаружи
@@ -112,10 +117,9 @@ const vClickOutside = {
 <style scoped>
 .select-box-main {
   display: flex;
-  width: 100%;
+  width: max-content;
   flex-direction: column;
   position: relative;
-  flex: 1;
 }
 
 .options-container-main {
@@ -216,8 +220,6 @@ const vClickOutside = {
   align-items: center;
   justify-content: space-between;
   padding-left: 5px;
-  word-break: break-all;
-
 
   cursor: pointer;
   padding: 5px 5px;
@@ -228,12 +230,6 @@ const vClickOutside = {
   position: relative;
 }
 
-@media (max-width: 620px) {
-  .options-container-main {
-    max-width: 80%; 
-    word-break: break-all;
-  }
-}
 
 .dark .selected-main {
   color: var(--white);
