+++ src/js/AuthFunctions.js
@@ -3,19 +3,11 @@ import { configData } from "./configData.js";
 export function isManager() {
   const managerDeviceName = localStorage.getItem(configData.MANAGER_DEVICE_NAME);
   const managerTokName = localStorage.getItem(configData.MANAGER_TOK_NAME);
-  const managerTypeName = localStorage.getItem(configData.MANAGER_TYPE_NAME);
 
   const deviceIsCorrect = managerDeviceName !== null && managerDeviceName !== undefined && managerDeviceName !== '';
   const tokenIsCorrect = managerTokName !== null && managerTokName !== undefined && managerTokName !== '';
-  const typeIsCorrect = managerTypeName !== null && managerTypeName !== undefined && managerTypeName !== '';
 
-  return (deviceIsCorrect && tokenIsCorrect && typeIsCorrect) ? true : false;
-}
-
-export function isAdmin() {
-  const managerTypeName = localStorage.getItem(configData.MANAGER_TYPE_NAME);
-
-  return isManager() && managerTypeName === 'Админ';
+  return (deviceIsCorrect && tokenIsCorrect) ? true : false;
 }
 
 export function isSeeker() {
@@ -34,7 +26,6 @@ export function logOut(whoIs) {
     case "manager":
       localStorage.removeItem(configData.MANAGER_TOK_NAME);
       localStorage.removeItem(configData.MANAGER_DEVICE_NAME);
-      localStorage.removeItem(configData.MANAGER_TYPE_NAME);
       break;
     case "seeker":
       localStorage.removeItem(configData.SEEKER_TOK_NAME);
