+++ src/js/configData.js
@@ -13,8 +13,6 @@ export let configData = {
     "263a4dbe41488fb87214b0032339dbb9f0c8da14c16dfcf13084bf3c2552eca0",
   MANAGER_TOK_NAME:
     "3c469e9d6c5875d37a43f353d4f88e61fcf812c66eee3457465a40b0da4153e0",
-  MANAGER_TYPE_NAME:
-    "MANAGER_TYPE_NAME",
   SEEKER_DEVICE_NAME:
     "263a4dbe41488fb87214b0032339dbb9f0c8da14c16dfcf13084bf3c2552eca1",
   SEEKER_TOK_NAME:
