+++ src/pages/vacancy/VacancyEdit/components/VacancyStatusList.vue
@@ -1,0 +1,331 @@
+<template>
+  <div class="vacancy-edit__statuslist">
+    <div>Статусы: {{ statusList.statuses.length || 'Статусы не заданы' }}</div>
+    <div class="statuslist__list-box">
+      <div
+        class="statuslist__list-items"
+        v-for="status in statusList.statuses"
+        :key="status.statusName"
+      >
+        <div
+          class="statuslist__list-status"
+          @click="
+            indicators.activeHandlers =
+              !indicators.activeHandlers ||
+              statusMod.name !== status.statusName;
+            statusMod.name = status.statusName;
+          "
+        >
+          <div
+            class="statuslist__list-comment"
+            :style="{ backgroundColor: status.statusColor }"
+          >
+            {{ status.statusName }}
+            <div
+              v-if="status.statusComment"
+              class="statuslist__list-comment-tooltip"
+            >
+              {{ status.statusComment || 'Нет комментария' }}
+            </div>
+            <ButtonIcon
+              class="statuslist__list-btn"
+              @click="
+                (indicators.isEdit = true),
+                  (statusMod.name = status.statusName),
+                  (statusMod.comment = status.statusComment),
+                  (statusMod.color = status.statusColor)
+              "
+            >
+              <template v-slot:icon
+                ><IconEdit
+                  :class="'statuslist__list-icon-edit'"
+                  :style="{
+                    display:
+                      indicators.activeHandlers &&
+                      statusMod.name === status.statusName
+                        ? 'block'
+                        : 'none',
+                  }"
+              /></template>
+            </ButtonIcon>
+            <ButtonIcon
+              class="statuslist__list-btn"
+              @click="handleModification(status.statusName, 'delete')"
+            >
+              <template v-slot:icon
+                ><IconDelete
+                  :class="'statuslist__list-icon-delete'"
+                  :style="{
+                    display:
+                      indicators.activeHandlers &&
+                      statusMod.name === status.statusName
+                        ? 'block'
+                        : 'none',
+                  }"
+              /></template>
+            </ButtonIcon>
+          </div>
+        </div>
+        <div v-if="statusList.transferStatuses[status.statusName].length">
+          <IconArrowSharp class="statuslist__list-arrow" />
+        </div>
+        <div
+          v-if="statusList.transferStatuses[status.statusName].length"
+          class="statuslist__list-transfers"
+        >
+          <div
+            class="statuslist__list-transfers-box"
+            v-for="transfer in statusList.transferStatuses[status.statusName]"
+          >
+            <div
+              class="statuslist__list-transfers-item"
+              :style="{
+                backgroundColor: statusList.statuses.find(
+                  (s) => s.statusName === transfer
+                ).statusColor,
+              }"
+              @click="
+                indicators.activeTransfer =
+                  !indicators.activeTransfer ||
+                  statusMod.name !== status.statusName;
+                statusMod.name = status.statusName;
+              "
+            >
+              {{ transfer || '' }}
+              <ButtonIcon
+                class="statuslist__list-btn"
+                @click="
+                  handleModification(
+                    status.statusName,
+                    'delete',
+                    true,
+                    transfer
+                  )
+                "
+              >
+                <template v-slot:icon
+                  ><IconDelete
+                    class="statuslist__list-icon-delete"
+                    :style="{
+                      display:
+                        indicators.activeTransfer &&
+                        statusMod.name === status.statusName
+                          ? 'block'
+                          : 'none',
+                    }"
+                /></template>
+              </ButtonIcon>
+            </div>
+          </div>
+        </div>
+        <ButtonIcon
+          class="statuslist__list-btn"
+          @click="
+            if (indicators.isTransfer && statusMod.name !== status.statusName) {
+              statusMod.name = status.statusName;
+            } else
+              (indicators.isTransfer = !indicators.isTransfer),
+                (statusMod.name = status.statusName);
+          "
+        >
+          <template v-slot:icon
+            ><IconAdd class="statuslist__list-icon"
+          /></template>
+        </ButtonIcon>
+        <SelectMain
+          v-if="indicators.isTransfer && statusMod.name === status.statusName"
+          class="statuslist__list-select"
+          v-model="statusMod.toName"
+          :options="
+            statusList.statusesSelect.filter(
+              (el) =>
+                el.id !== status.statusName &&
+                statusList.transferStatuses[status.statusName].indexOf(
+                  el.id
+                ) === -1
+            )
+          "
+          @update:model-value="
+            handleModification(status.statusName, 'create', true)
+          "
+        />
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup>
+import SelectMain from '@/components/SelectMain.vue';
+import ButtonIcon from '@/components/ButtonIcon.vue';
+import IconAdd from '@/assets/icons/add.svg?component';
+import IconEdit from '@/assets/icons/edit.svg?component';
+import IconDelete from '@/assets/icons/close.svg?component';
+import IconArrowSharp from '@/assets/icons/arrow-sharp.svg?component';
+const props = defineProps({
+  // Массив статусов
+  statusList: {
+    type: [Array, Object, Array],
+    required: true,
+  },
+  // Информация об изменяемом статусе
+  statusMod: {
+    type: Object,
+    required: true,
+  },
+  // Индикаторы кнопок
+  indicators: {
+    type: Object,
+    required: true,
+  },
+  // Обработчик изменения статуса
+  handleModification: {
+    type: Function,
+    required: true,
+  },
+});
+</script>
+
+<style scoped>
+.vacancy-edit__statuslist {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.statuslist__list-arrow {
+  height: fit-content;
+  display: flex;
+  fill: black;
+  width: 25px;
+  height: 25px;
+}
+
+.statuslist__list-box {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.statuslist__list-items {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  user-select: none;
+}
+
+.statuslist__list-color {
+  width: 10px;
+  height: 10px;
+  border-radius: 100%;
+}
+
+.statuslist__list-status {
+  display: flex;
+  cursor: pointer;
+}
+
+.statuslist__list-transfers {
+  display: flex;
+  gap: 5px;
+  cursor: pointer;
+}
+
+.statuslist__list-transfers-box {
+  display: flex;
+  align-items: end;
+  &::after {
+    content: ',';
+  }
+  &:last-child::after {
+    content: '';
+  }
+}
+
+.statuslist__list-transfers-item {
+  display: flex;
+  padding: 5px;
+  border-radius: 10px;
+  color: white;
+  padding: 7px 14px;
+}
+
+.statuslist__list-comment-tooltip {
+  visibility: hidden;
+  position: absolute;
+  z-index: 1;
+  margin: 0;
+  margin-top: -70px;
+  margin-left: -3px;
+  max-width: 200px;
+  max-height: 100px;
+  display: block;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  word-break: break-all;
+  text-overflow: ellipsis;
+  background-color: var(--cornflower-blue);
+  color: var(--white);
+  text-align: center;
+  border-radius: 6px;
+  padding: 5px 10px;
+  font-size: 12px;
+}
+
+.statuslist__list-comment {
+  min-height: 25px;
+  display: flex;
+  align-items: center;
+  color: white;
+  border-radius: 10px;
+  padding: 5px 12px;
+}
+
+.statuslist__list-comment:hover,
+.statuslist__list-transfers-item:hover {
+  .statuslist__list-comment-tooltip {
+    visibility: visible;
+  }
+}
+
+.statuslist__list-btn {
+  display: flex;
+  padding: 0;
+  width: fit-content;
+}
+
+.statuslist__list-icon {
+  align-self: center;
+  transition: all 0.3s ease;
+  width: 20px;
+  height: 20px;
+  &:hover {
+    transform: scale(1.3);
+  }
+}
+
+.statuslist__list-select {
+  font-size: 13px;
+  height: 20px;
+}
+
+.statuslist__list-icon-edit,
+.statuslist__list-icon-delete {
+  display: none;
+  position: relative;
+  margin: 0;
+  margin-top: -15px;
+  height: 12px;
+  width: 12px;
+  fill: white;
+  &.active {
+    display: block;
+  }
+}
+
+.statuslist__list-icon-edit {
+  margin-left: 5px;
+}
+</style>
